Start of SubAccounts
Allows for creation of sub accounts and modifications of existing code
base to account for sub accounts. Queries to the DB require both the
current user and the current sub account values to retrieve the correct
user-sub account pair.

There is still a lot of work and fine tuning to finish sub accounts.